requirements.txt
Markdown==3.9
Pillow==12.0.0
pyside6==6.9.2
pyside6_addons==6.9.2
pyside6_essentials==6.9.2
Requests==2.32.5
torch==2.8.0
packaging==25.0
tqdm==4.67.1
torchvision==0.24.0